% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mishmash.R
\name{flash_get_E_pm}
\alias{flash_get_E_pm}
\title{Get posterior mean of E from flash model}
\usage{
flash_get_E_pm(fl, fit_E)
}
\arguments{
\item{fl}{A \code{flash} object.}

\item{fit_E}{Boolean indicating if E was fit in the flash model. This is
equivalent to \code{var.type = NULL} in \code{flashier::flash}.}
}
\value{

}
\description{
Get posterior mean of E from flash model
}
