% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mishmash.R
\name{flash_get_pm}
\alias{flash_get_pm}
\title{Get Posterior Mean of Flash Model}
\usage{
flash_get_pm(fl)
}
\arguments{
\item{fl}{flash object}
}
\value{
matrix with posterior mean of each element of data matrix
}
\description{
Computes posterior mean of LF' + E. Note that this will in general
give different results than \code{flashier::flash}, as that function doesn't
add the posterior mean of E.
}
\examples{

# Generate 10 x 500 matrix with rank 3.
n <- 10
p <- 500
k <- 3
A <- matrix(
   data = rnorm(n = n * k, sd = 2), nrow = n, ncol = k
)

cov <- A \%*\% t(A) / (n - 1)

X <- t(MASS::mvrnorm(n = p, mu = rep(0, n), Sigma = cov))
Y <- X + matrix(rnorm(n * p, 0, 1), n, p)

# set mash data
mash_data <- mashr::mash_set_data(Y)

# call flash
fl <- mashy_flash(mash_data, backfit = TRUE)

# get posterior mean
pm <- flash_get_pm(fl)

}
