% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mishmash.R
\name{flash_sample_fx}
\alias{flash_sample_fx}
\title{Sample Posterior Effect Sizes from Flash Model}
\usage{
flash_sample_fx(flash, nsamp)
}
\arguments{
\item{flash}{flash object}

\item{nsamp}{number of samples}
}
\value{
List of samples
}
\description{
Sample Posterior Effect Sizes from Flash Model
}
