% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_writer.R
\name{lazy_writer}
\alias{lazy_writer}
\title{convenience wrapper to save off object as cvs and RDS}
\usage{
lazy_writer(rds_output_directory = NULL, csv_output_directory = NULL)
}
\arguments{
\item{rds_output_directory}{the directory to save the RDS file to}

\item{csv_output_directory}{the directory to save the csv file to}
}
\value{
a \code{function} which in turn writes out an object to a csv and RDS file
}
\description{
A convenience closure to wrap around write_out
}
\details{
This function creates a function which will save off RDS and csv files to
named directories
}
\examples{
\dontrun{
# creates a function which will save RDS and csv files to the
# "rds_files" and  "csv_files" directories
x <- lazy_writer("rds_files","csv_files")

# write out the files
x(mtcars)
}
}
