% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mishmash.R
\name{mashy_flash}
\alias{mashy_flash}
\title{Run Flash with mash data object}
\usage{
mashy_flash(mash_data, lfsr_nsamp = 1000, ...)
}
\arguments{
\item{mash_data}{Output from \code{mashr::mash_set_data}.}

\item{lfsr_nsamp}{number of samples for computation of lfsr. Set to 0
to skip computation of lfsr.}

\item{...}{Additional arguments to be passed to \code{flashier::flash}.}
}
\value{
A \code{flash} object. For more details see the documentation for
\code{flashier::flash}.
}
\description{
Run Flash with mash data object
}
\examples{

# Generate 10 x 500 matrix with rank 3.
n <- 10
p <- 500
k <- 3
A <- matrix(
   data = rnorm(n = n * k, sd = 2), nrow = n, ncol = k
)

cov <- A \%*\% t(A) / (n - 1)

X <- t(MASS::mvrnorm(n = p, mu = rep(0, n), Sigma = cov))
Y <- X + matrix(rnorm(n * p, 0, 1), n, p)

# set mash data
mash_data <- mashr::mash_set_data(Y)

# call flash
fl <- mashy_flash(mash_data, backfit = TRUE)

}
