% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mishMashr.R
\docType{package}
\name{mishMashr}
\alias{mishMashr}
\alias{mishMashr-package}
\title{Little helpers to make life easier...}
\description{
A collection of general helper functions to make it easier to look for data
quality issues and some helpers for saleforce analysis
}
\author{
Mark Goble \email{mark.goble@goble.co.uk}
}
