% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_group_check.R
\name{mixed_group_check}
\alias{mixed_group_check}
\title{Find parents with two or more different child values in the child column}
\usage{
mixed_group_check(data_frame, parent_column, child_column)
}
\arguments{
\item{data_frame}{a data frame to check}

\item{parent_column}{the parent column on the data frame}

\item{child_column}{the child column to check}
}
\value{
A \code{data frame} containing parents where there is more than 1 child value and the number of occurances.
}
\description{
Produces a list of parents which have more than one distinct child value and the number of
children. It will filter out any parents with only one child value.
This is useful in finding issues where the same parent has different values on the children :
for example where all products should have the same colour description
}
\examples{
mixed_group_check(mtcars,"vs","cyl")
}
