% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_group_check2.R
\name{mixed_group_check2}
\alias{mixed_group_check2}
\title{Finds parents with two or more child values in two children columns}
\usage{
mixed_group_check2(data_frame, parent_column, child_column_1,
  child_column_2)
}
\arguments{
\item{data_frame}{a data frame to check}

\item{parent_column}{the parent column on the data frame}

\item{child_column_1}{the first child column to check}

\item{child_column_2}{the second child column to check}
}
\value{
A \code{data frame} containing parents where there is more than 1 child value and the number of occurances.
}
\description{
Produces a list of parents which have more than one distinct child value and the number of
children taking two child inputs. It will filter out any parents with only one child value.
This is useful in finding issues where the same parent has different values on the children :
for example where all products should have the same colour description
}
\examples{
mixed_group_check2 (mtcars,"vs","cyl", "am")
}
