% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/once_only.R
\name{once_only}
\alias{once_only}
\title{Efficiently find rows where a the value appears only once in the column}
\usage{
once_only(data_frame, column, return_column_only = TRUE)
}
\arguments{
\item{data_frame}{a data frame to read in}

\item{column}{the column in the data frame to check}

\item{return_column_only}{should the function return the whole data frame or just the column?}
}
\value{
A \code{dataframe} of values from the column which aren't duplicated.
}
\description{
Returns a data frame of values which only appear once in a df column.
}
\details{
Returns only the rows from a data frame if there is only one instance of the
value in the specified column. This method is more efficient than grouping by
the column and then filtering for rows with a count == 1.

Setting the \code{return_column_only = TRUE} will return the data frame with
all of the instances where the value of column appears more than once removed.
}
\examples{
mtcars$model <- rownames(mtcars)
# just the values from the column which appear only once
once_only(mtcars,"mpg")
# the rows for the values of the column appears once only
y <- once_only(mtcars,"mpg", return_column_only = FALSE )
}
