% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactorer.R
\name{refactorer}
\alias{refactorer}
\title{easier way to refactor using lists}
\usage{
refactorer(
  data_frame,
  column,
  level_vector,
  label_vector,
  other_label = "Unknown"
)
}
\arguments{
\item{data_frame}{The data frame containing the column}

\item{column}{The column to refactor}

\item{level_vector}{The levels in the orignal column}

\item{label_vector}{The labels to apply}

\item{other_label}{The label to use for 'other' values (e.g. unknowns)}
}
\value{
an updated column
}
\description{
Refactor a factor using input lists
}
\details{
This function allows for quick bulk replacement of the labels of a factor
rather than the more complex forcats way...
}
\examples{
\dontrun{
To add!
}
}
