% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_html.R
\name{remove_html}
\alias{remove_html}
\title{Strip HTML from text}
\usage{
remove_html(string)
}
\arguments{
\item{string}{A string to clean}
}
\value{
A cleaned \code{string}
}
\description{
Remove HTML Tags from a string
}
\details{
Remove all the HTML tags from a string input
}
\examples{
\dontrun{
Add examples
}
}
