% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_connection.R
\name{secure_connection}
\alias{secure_connection}
\title{securely connect to a data source using a connection file}
\usage{
secure_connection(connection_file = NULL, directory = NULL)
}
\arguments{
\item{connection_file}{an r script file containing the connection}

\item{directory}{the directory containing the connection file}
}
\description{
Source the connection details from a connection file. Updated to allow passing
in of the connection file details including the directory.
Note this version uses the rstudioapi package so intended to be used within
rstudio
}
\details{
Source the connection details from a connection file
}
\examples{
\dontrun{
Insert example here
}
}
