% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling_folder.R
\name{sibling_folder}
\alias{sibling_folder}
\title{returns the full path for a sibling folder}
\usage{
sibling_folder(destination_folder)
}
\arguments{
\item{destination_folder}{the desination folder you want to reach (in quotes)}
}
\value{
the full path to the destination folder
}
\description{
A quick and easy way to get a sibling folder
}
\details{
A quick function to access a sibling folder (i.e. a folder with the same parent folder).
}
\examples{
\dontrun{
To add!
}
}
