% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_loader.R
\name{source_loader}
\alias{source_loader}
\title{safely loads source file r scripts}
\usage{
source_loader(source_file, path)
}
\arguments{
\item{source_file}{the R script to source}

\item{path}{the path to the file}
}
\value{
none
}
\description{
A wrapper around source
}
\details{
A wrapper around source. Recommended to use with the \link[marksLittleHelper]{sibling_folder} function.
}
\examples{
\dontrun{
To add!
}
}
