% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_out.R
\name{write_out}
\alias{write_out}
\title{consistently write out data files}
\usage{
write_out(
  object,
  add_date = FALSE,
  output_directory = NULL,
  file_type = "RDS",
  custom_name = NULL
)
}
\arguments{
\item{object}{the object to be saved}

\item{add_date}{should the date be added to the end of the file name?}

\item{output_directory}{should the output be saved to a particular directory (default to working directory)}

\item{file_type}{should the file be saved as RDS or csv (default RDS)}

\item{custom_name}{should a custom name be used instead of the objects' name}
}
\value{
none
}
\description{
A wrapper around write.csv and saveRDS
}
\details{
This function provides consistent saving. By default the file will be saved
with the name of the name of the object.

The format of the date is _(day)_(month)_(year)
}
\examples{
\dontrun{
 write_out(mtcars)
 # this will save the mtcars data to an RDS file.

 write_out(mtcars, add_date = TRUE, file_type = "csv", custom_name = "cars")
 # writes the mtcars data to a csv file "cars_(day)_(month)_(year).csv"

}
}
