% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling_folder.R
\name{sibling_folder}
\alias{sibling_folder}
\title{returns the full path for a sibling folder}
\usage{
sibling_folder(destination_folder, parent_level = FALSE)
}
\arguments{
\item{destination_folder}{the destination folder you want to reach (in quotes)}

\item{parent_level}{if true then return the sibling of the parent folder (default = false)}
}
\value{
the full path to the destination folder
}
\description{
A quick and easy way to get a sibling folder
}
\details{
A quick function to access a sibling folder (i.e. a folder with the same parent folder).
If parent_level is set as FALSE (defaults to false) it will look on the same level in
the directory structure; however if parent_level is set to TRUE then it will look for the
sibling folder at the parent level (I hope that makes sense!)
}
\examples{
\dontrun{
To add!
}
}
